extract api method
